Refactor crew.py and document backend structure
This commit refactors the `crew.py` script by moving it to the `utils` directory and separating the LLM configuration into a new `core/llm.py` module. It also moves the `generate_final_markdown` function to the `utils/doc_export.py` module.

Additionally, this commit adds a `README.md` file to the `backend` directory that documents the new, modular code structure.
