+++ backend/utils/crew.py
@@ -7,55 +7,7 @@ import os
 import uuid
 from datetime import datetime
 
-from langchain_openai import AzureChatOpenAI
-from dotenv import load_dotenv
-
-# Load environment variables
-load_dotenv()
-
-# Configure Azure OpenAI
-os.environ["AZURE_API_TYPE"] = "azure"
-os.environ["AZURE_API_BASE"] = os.getenv("AZURE_OPENAI_ENDPOINT")
-os.environ["AZURE_API_KEY"] = os.getenv("AZURE_OPENAI_API_KEY")
-os.environ["AZURE_API_VERSION"] = os.getenv("OPENAI_API_VERSION")
-os.environ["AZURE_DEPLOYMENT_NAME"] = os.getenv("AZURE_DEPLOYMENT_NAME")
-# os.environ["EMBEDDING_MODEL"] = os.getenv("EMBEDDING_MODEL")
-
-# Validate environment variables
-required_vars = ["AZURE_OPENAI_ENDPOINT", "AZURE_OPENAI_API_KEY", "OPENAI_API_VERSION", "AZURE_DEPLOYMENT_NAME"]
-missing_vars = [var for var in required_vars if not os.getenv(var)]
-if missing_vars:
-    raise ValueError(f"Missing required environment variables: {', '.join(missing_vars)}")
-
-# Initialize LLM
-llm = AzureChatOpenAI(
-    azure_deployment=os.getenv("AZURE_DEPLOYMENT_NAME"),
-    azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"),
-    api_key=os.getenv("AZURE_OPENAI_API_KEY"),
-    api_version=os.getenv("OPENAI_API_VERSION"),
-    model=f"azure/{os.getenv('AZURE_DEPLOYMENT_NAME')}",
-    # deployment_model=f"azure/{os.getenv('"EMBEDDING_MODEL"')}",
-    max_retries=3,
-    timeout=30
-)
-
-
-## Atlernative with Google 
-#from langchain_google_genai import ChatGoogleGenerativeAI
-#llm2 = ChatGoogleGenerativeAI(
-#    model="gemini-1.5-flash",  # Or "gemini-1.5-pro", or #other available Gemini models
-#    verbose=True, # For seeing more detailed output from #the LLM
-#    temperature=0.7, # Adjust as needed
-#    google_api_key=os.getenv("GEMINI_API_KEY") # Or #GOOGLE_API_KEY
-#)
-#from crewai import LLM
-#llm3 = LLM(
-#     model="gemini/gemini-1.5-flash", # Or "gemini/#gemini-1.5-pro"
-#     temperature=0.7,
-#     api_key=os.getenv("GEMINI_API_KEY") # 
-# )
-
-# The proposal data is now loaded in backend.core.config
+from backend.core.llm import llm
 
 @CrewBase
 class ProposalCrew():
@@ -178,33 +130,5 @@ class ProposalCrew():
         )
    
 
-    # Move this function to Tools folder
-    def generate_final_markdown(self, generated_sections):
-        """Compile all generated sections into a markdown file with a unique filename."""
-
-        # ✅ Define a fixed folder name
-        folder_name = "proposal-documents"
-        folder_path = os.path.join(os.getcwd(), folder_name)
-
-        # ✅ Create the folder if it doesn’t exist
-        os.makedirs(folder_path, exist_ok=True)
-
-        # ✅ Generate a unique ID for the file (timestamp + random 6-char UUID)
-        unique_id = f"{datetime.now().strftime('%Y%m%d%H%M%S')}_{uuid.uuid4().hex[:6]}"
-        file_name = f"proposal_document_{unique_id}.md"
-
-        # ✅ Define file path inside the fixed folder
-        file_path = os.path.join(folder_path, file_name)
-
-        # ✅ Generate proposal content
-        markdown_content = ""
-        for section, content in generated_sections.items():
-            markdown_content += f"{section}\n{content.strip()}\n\n"
-
-        # ✅ Save the file with a unique name inside the `proposal-documents/` folder
-        with open(file_path, "w", encoding="utf-8") as md_file:
-            md_file.write(markdown_content)
-
-        print(f"✅ Proposal document saved at: {file_path}")
